Moves blog posts list component to PostsSection from PostsPage.
