+++ app/posts/page.tsx
@@ -1,60 +1,13 @@
-import { DataList, Grid, Text } from "@radix-ui/themes";
-import { Suspense } from "react";
-import AccentedHeading from "../components/AccentedHeading";
-import { WpPost, fetchWpPosts } from "../lib/server-lib";
-import { cutOffText } from "../lib/utils";
+import LandingSection from "../contact/sections/LandingSection";
+import PostsSection from "./sections/PostsSection";
 
-export const revalidate = 3600; // 1 hour
-
-const PlaceholderPost = ({
-  text = "Loading blog posts..",
-  title = "Loading...",
-}: {
-  text?: string;
-  title?: string;
-}) => (
-  <DataList.Item>
-    <DataList.Label>{title}</DataList.Label>
-    <DataList.Value>
-      <Text as="p">{text}</Text>
-    </DataList.Value>
-  </DataList.Item>
-);
-
-export default async () => {
-  const posts: WpPost[] | 1 = await fetchWpPosts();
-
+const PostsPage: React.FC = () => {
   return (
-    <article>
-      <AccentedHeading
-        textAs="h1"
-        size="9"
-        preText="My "
-        accentedText="Dev Blog."
-      />
-      <DataList.Root>
-        <Suspense fallback={<PlaceholderPost text="Fetching blog posts..." />}>
-          {posts === 1 ? (
-            <PlaceholderPost
-              title="Error!"
-              text="Error while fetching blog posts... Please try again later."
-            />
-          ) : (
-            posts.map(({ id, excerpt, date, title }) => (
-              <DataList.Item key={id}>
-                <DataList.Label>{title.rendered}</DataList.Label>
-                <DataList.Value>
-                  <Grid columns={{ sm: "2", md: "3" }}>
-                    <Text as="p">{id}</Text>
-                    <Text as="p">{new Date(date).toLocaleDateString()}</Text>
-                    <Text as="p">{cutOffText(excerpt.rendered, 100)}</Text>
-                  </Grid>
-                </DataList.Value>
-              </DataList.Item>
-            ))
-          )}
-        </Suspense>
-      </DataList.Root>
-    </article>
+    <>
+      <LandingSection />
+      <PostsSection />
+    </>
   );
 };
+
+export default PostsPage;
